backend/requirementsB.txt
fastapi
uvicorn
pandas
numpy==1.26.4
scikit-surprise
requests
